Classes que compõem hotel e seus respectivos testes Foram criadas as classes Hotel, Hóspede, Contrato e Opinião; Foi criado um enum básico: EstadoDeContrato; Testes de cada classe foram feitos; Ajustes para integrar o sistema às classes novas;

+++ src/alugaveis/TipoQuarto.java
@@ -12,17 +12,17 @@ package alugaveis;
  */
 public enum TipoQuarto {
 
-	PRESIDENCIAL(1200.0, 4, "Quarto Presidencial\nQuarto equipado com TV LCD 42, split, frigobar, cofre,sala de jogos e home theater, ideal para familias em ferias.\n"),
+	PRESIDENCIAL(1200, 4, "Quarto Presidencial\nQuarto equipado com TV LCD 42, split, frigobar, cofre,sala de jogos e home theater, ideal para familias em ferias.\n"),
 	
-	LUXO_SIMPLES(520.0, 1, "Quarto Luxo Simples\nQuarto equipado com TV LCD 42, split, frigobar, cofre e home theater.\n"),
-	LUXO_DUPLO(570.0, 2, "Quarto Luxo Duplo\nQuarto equipado com TV LCD 42, split, frigobar, cofre e home theater.\n"),
-	LUXO_TRIPLO(620.0, 3, "Quarto Luxo Triplo\nQuarto equipado com TV LCD 42, split, frigobar, cofre e home theater.\n"),
+	LUXO_SIMPLES(520, 1, "Quarto Luxo Simples\nQuarto equipado com TV LCD 42, split, frigobar, cofre e home theater.\n"),
+	LUXO_DUPLO(570, 2, "Quarto Luxo Duplo\nQuarto equipado com TV LCD 42, split, frigobar, cofre e home theater.\n"),
+	LUXO_TRIPLO(620, 3, "Quarto Luxo Triplo\nQuarto equipado com TV LCD 42, split, frigobar, cofre e home theater.\n"),
 	
-	EXECUTIVO_SIMPLES(360.0, 1, "Quarto Executivo Simples\nQuarto equipado com TV LCD 42, split, frigobar e cofre.\n"),
-	EXECUTIVO_DUPLO(385.0, 2, "Quarto Executivo Duplo\nQuarto equipado com TV LCD 42, split, frigobar e cofre.\n"),
-	EXECUTIVO_TRIPLO(440.0, 3 , "Quarto Executivo Triplo\nQuarto equipado com TV LCD 42, split, frigobar e cofre.\n");
+	EXECUTIVO_SIMPLES(360, 1, "Quarto Executivo Simples\nQuarto equipado com TV LCD 42, split, frigobar e cofre.\n"),
+	EXECUTIVO_DUPLO(385, 2, "Quarto Executivo Duplo\nQuarto equipado com TV LCD 42, split, frigobar e cofre.\n"),
+	EXECUTIVO_TRIPLO(440, 3 , "Quarto Executivo Triplo\nQuarto equipado com TV LCD 42, split, frigobar e cofre.\n");
 	
-	private double diaria;
+	private float diaria;
 	private int capacidade;
 	private String descricao;
 	
@@ -35,7 +35,7 @@ public enum TipoQuarto {
 	 * @param descricao
 	 * 			descricao do quarto.
 	 */
-	TipoQuarto(double diaria, int capacidade, String descricao){
+	TipoQuarto(float diaria, int capacidade, String descricao){
 		this.diaria = diaria;
 		this.capacidade = capacidade;
 		this.descricao = descricao;
@@ -46,7 +46,7 @@ public enum TipoQuarto {
 	 * @return
 	 * 			Valor da diaria.
 	 */
-	public double getDiaria() {
+	public float getDiaria() {
 		return diaria;
 	}
